% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BG_test.R
\name{BG_test}
\alias{BG_test}
\title{Breusch-Godfrey autocorrelation test}
\usage{
BG_test(object, lag, type = c("LR", "LM", "Rao"))
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{lag}{The lag order of auxiliary regression model.}

\item{type}{The test version. It can be selected between \code{LR} (the default), \code{LM} and \code{Rao}.}
}
\value{
An object of class \code{BG_test} is a list with the following components: \cr

\item{stat}{The test statistic.}
\item{stat_pval}{The p-value of the test statistic.}
\item{stat_adj}{The adjusted test statistic (not available for \code{Rao}).}
\item{stat_adj_pval}{The p-value of the adjusted test statistic (not available for \code{Rao}).}
\item{df}{The degrees of freedom of the test statistic.}
\item{type}{The version of the test.}
\item{lag}{The lag order of auxiliary regression model.}
}
\description{
This function tests the serial correlation of residuals using the Breusch-Godfrey test. The user can choose between three versions of the test: likelihood ratio (\code{LR}), Lagrange multiplier (\code{LM}) and Rao.
}
\details{
This test should be applied for a \code{lag} less than the VECM lag order. For a greater lag, it is suggested to use the portmanteau test (\code{\link{port_test}}).
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)
test_LR <- BG_test(object = model, lag = 2, type = "LR")
test_Rao <- BG_test(object = model, lag = 4, type = "Rao")

print(test_LR)
print(test_Rao)
}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
