% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chow_test.R
\name{chow_test}
\alias{chow_test}
\title{Chow structural break test}
\usage{
chow_test(object, t_break)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{t_break}{An integer. The period for which the structural break is tested.}
}
\value{
An object of class \code{chow_test} is a list with the following components: \cr

\item{stat_SS}{The sample-split test statistic.}
\item{df_SS}{The degrees of freedom of the test statistic \code{stat_SS}.}
\item{pval_SS}{The p-value of the test statistic \code{stat_SS}.}
\item{stat_BP}{The break-point test statistic.}
\item{df_BP}{The degrees of freedom of the test statistic \code{stat_BP}.}
\item{pval_BP}{The p-value of the test statistic \code{stat_BP}.}
\item{stat_F}{The forecast test statistic.}
\item{df_F}{The degrees of freedom of the test statistic \code{stat_F}.}
\item{pval_F}{The p-value of the test statistic \code{stat_F}.}
}
\description{
This function tests the presence of a structural change by implementing the sample-split (\code{SS}), the break-point (\code{BP}) and the forecast (\code{F}) versions of the Chow test.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

chow <- chow_test(object = model, t_break = 43)
print(chow)
}

}
\references{
Doornik, J. A. & Hendry, D. F. (1997). \emph{Modelling Dynamic Systems Using PcFiml 9.0 for Windows}, International Thomson Business Press, London.

L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
