% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.VECM.R
\name{coef.VECM}
\alias{coef.VECM}
\title{Extract VECM coefficients}
\usage{
\method{coef}{VECM}(x)
}
\arguments{
\item{x}{An object of class \code{VECM}.}
}
\value{
A list with the following components: \cr

\item{alpha}{The loading matrix.}
\item{beta}{The cointegrating matrix.}
\item{pi}{The long-run impact matrix.}
\item{gamma}{The short-run impact matrix.}
\item{D_ect}{The matrix containing the estimated deterministic terms within the error correction term.}
\item{D}{The matrix containing the estimated deterministic terms.}
}
\description{
This function is a method for the class \code{VECM}.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
