% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointegration_test.R
\name{cointegration_test}
\alias{cointegration_test}
\title{Cointegration test}
\usage{
cointegration_test(
  data,
  p,
  spec = c("none", "rconst", "uconst", "rtrend", "utrend"),
  method = c("johansen", "info_criteria"),
  alpha = 0.05,
  shift = NULL,
  season = NULL,
  season_ref = NULL,
  exogen_ect = NULL,
  exogen = NULL
)
}
\arguments{
\item{data}{A matrix containing the multivariate time series to be analyzed.}

\item{p}{The lag order.}

\item{spec}{The type of deterministic component to be considered. It can be selected between \code{none} (the default), \code{rconst}, \code{uconst}, \code{rtrend}, and \code{utrend}.}

\item{method}{The estimation method of the cointegrating rank. It can be selected between \code{johansen} (the default) and \code{info_criteria}.}

\item{alpha}{The significance level used for hypothesis testing.}

\item{shift}{An integer or a vector. The break point at which there is a change in the levels of the variables.}

\item{season}{The number of seasonal dummy variables.}

\item{season_ref}{The season selected as reference. If \code{NULL}, the last season is automatically used as reference.}

\item{exogen_ect}{The exogenous variable to be included in the error correction term.}

\item{exogen}{The exogenous variable to be included outside the error correction term.}
}
\value{
An object of class \code{cointegration_test} is a list with the following components: \cr

\item{n}{The number of observations.}
\item{p}{The lag order.}
\item{spec}{The type of deterministic component.}
\item{method}{The estimation method of the cointegrating rank.}
\item{n_par}{A vector containing the number of estimated parameters.}
\item{e_val}{A vector cotaining the eigenvalues used to implement the test.}
\item{stat}{A matrix containing the Johansen test statitics.}
\item{crit_val}{A matrix containing the critical values associated with the Johansen distribution.}
\item{p_val}{A matrix containing the p-values of the Johansen test statistics.}
\item{alpha}{The probability level at which the critical values are computed.}
\item{log_lik}{A vector containing the log-likelihood of the estimated VEC models.}
\item{IC}{A matrix containing the information criteria based on \code{log_lik}.}
}
\description{
This function implements a test to estimate the cointegrating rank between variables. The user can choose between the Johansen test and an information criteria-based test.
}
\details{
When choosing \code{method} = "johansen", the function incorporates both the trace and maximum eigenvalues tests. The determination of the cointegrating rank should consider the results from both tests.

When choosing \code{method} = "info_criteria", the function computes the most used information criteria, i.e. AIC, BIC, and HQC.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from Lütkepohl, 2005.
data("e6")

co_test <- cointegration_test(data = e6, p = 3, spec = "uconst", method = "johansen")

print(co_test)

}

}
\references{
Johansen, S. (1995). \emph{Likelihood-based Inference in Cointegrated Vector Autoregressive Models}. Oxford: Oxford University Press.

Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{johansen_distr}} \code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
