% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_plot.R
\name{ECT_plot}
\alias{ECT_plot}
\title{ECT_plot}
\usage{
ECT_plot()
}
\description{
Auxiliary function used in \code{\link{plot.VECM}}.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
\keyword{internal}
