% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger_test.R
\name{granger_test}
\alias{granger_test}
\title{Granger-causality test}
\usage{
granger_test(object, dip, indip)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{dip}{A character vector containing the names of the dependent variables.}

\item{indip}{A character vector containing the names of the independent variables.}
}
\value{
An object of class \code{granger_test} is a list with the following components: \cr

\item{R}{The matrix that identifies the parameters to be tested.}
\item{stat}{The test statistic.}
\item{stat_pval}{The p-value of the test statistic.}
\item{df}{The degrees of freedom of the test statistic.}
\item{dip}{A character vector containing the names of the dependent variables.}
\item{indip}{A character vector containing the names of the independent variables.}
}
\description{
This function implements the Granger-causality test for a VEC model.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

test_G <- granger_test(object = model, dip = "R", indip = "Dp")
print(test_G)
}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
