% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johansen_distr.R
\name{johansen_distr}
\alias{johansen_distr}
\title{Johansen distribution}
\usage{
johansen_distr(
  n_sim,
  n_step,
  rw_dim,
  spec = c("none", "rconst", "uconst", "rtrend", "utrend"),
  seed = NULL
)
}
\arguments{
\item{n_sim}{The number of simulations.}

\item{n_step}{The number of steps of a Gaussian random walk.}

\item{rw_dim}{The dimension of the random walk used.}

\item{spec}{The type of deterministic component to be considered. It can be selected between \code{none} (the default), \code{rconst}, \code{uconst}, \code{rtrend}, and \code{utrend}.}

\item{seed}{An optional argument specifying the seed for the simulation.}
}
\value{
A matrix containing the simulated values for the trace and maximum eigenvalue test statistics.
}
\description{
This function simulates the Johansen distribution for the trace and the maximum eigenvalue test statistics.
}
\references{
Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{johansen_stat}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
