% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johansen_stat.R
\name{johansen_stat}
\alias{johansen_stat}
\title{Johansen test statistics}
\usage{
johansen_stat(n_step, rw_dim, spec)
}
\arguments{
\item{n_step}{The number of steps of a Gaussian random walk.}

\item{rw_dim}{The dimension of the random walk used.}

\item{spec}{The type of deterministic component to be considered. It can be selected between \code{none} (the default), \code{rconst}, \code{uconst}, \code{rtrend}, and \code{utrend}.}
}
\value{
The values of the trace and the maximum eigenvalue test statistics.
}
\description{
This function computes the trace and maximum eigenvalue test statistics.
}
\references{
Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
