% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{normality_test}
\alias{normality_test}
\title{Multivariate normality tests}
\usage{
normality_test(object, type = c("Lutkepohl", "Doornik-Hansen", "Urzua"))
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{type}{The factorization of the residuals covariance matrix. It can be selected between \code{Lutkepohl} (the default), \code{Doornik-Hansen}, and \code{Urzua}.}
}
\value{
An object of class \code{VECM} is a list with the following components: \cr

\item{b1}{The estimated skewnesses of the residuals.}
\item{b1_chisq}{The skewness test statsitics.}
\item{b1_pval}{The p-value of the test statistic \code{b1_chisq}.}
\item{b1_joint}{The joint skewness test statsitic.}
\item{b1_joint_pval}{The p-value of the test statistic \code{b1_joint}.}
\item{b2}{The estimated kurtosis of the residuals.}
\item{b2_chisq}{The kurtosis test statsitic.}
\item{b2_pval}{The p-value of the test statistic \code{b2_chisq}.}
\item{b2_joint}{The joint kurtosis test statsitic.}
\item{b2_joint_pval}{The p-value of the test statistic \code{b2_joint}.}
\item{JB}{The Jarque-Bera test statsitic.}
\item{JB_pval}{The p-value of the test statistic \code{JB}.}
\item{JB_joint}{The joint Jarque-Bera test statsitic.}
\item{JB_joint_pval}{The p-value of the test statistic \code{JB_joint}.}
\item{type}{The factorization of the residuals covariance matrix.}
}
\description{
This function performs a series of normality tests for the residuals of a VEC model.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from Lütkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)
test <- normality_test(object = model, type = "Doornik-Hansen")

print(test)
}

}
\references{
Urzua, C. M. (1997). \emph{Omnibus tests for multivariate normality based on a class of maximum entropy distributions}. In Applying Maximum Entropy to Econometric Problems. Emerald Group Publishing Limited.

Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.

Doornik, J. A., & Hansen, H. (2008). \emph{An omnibus test for univariate and multivariate normality}. Oxford bulletin of economics and statistics, 70, 927-939.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
