% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orth_compl.R
\name{orth_compl}
\alias{orth_compl}
\title{Orthogonal complement}
\usage{
orth_compl(mat, r)
}
\arguments{
\item{mat}{The matrix for which the orthogonal complement is calculated.}

\item{r}{The rank of the matrix \code{mat}.}
}
\description{
Internal function used to compute the orthogonal complement of a matrix.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
\keyword{internal}
