% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VECM.R
\name{plot.VECM}
\alias{plot.VECM}
\title{Plotting VEC model fits}
\usage{
\method{plot}{VECM}(
  x,
  p_max = 25,
  alpha = 0.05,
  cross = FALSE,
  kernel = c("biweight", "epanechnikov", "gaussian", "rectangular", "triangular")
)
}
\arguments{
\item{x}{An object of class \code{VECM}.}

\item{p_max}{The maximum lag used in the calculation of the (partial) autocorrelation function.}

\item{alpha}{The confidence level used in the calculation of the confidence intervals.}

\item{cross}{A logical value. If \code{TRUE} (not the default), cross correlations are considered.}

\item{kernel}{A character string giving the smoothing kernel used in the calculation of the Kernel density estimates. It can be selected between \code{biweight} (the default), \code{epanechnikov}, \code{gaussian}, \code{rectangular}, and \code{triangular}.}
}
\description{
This function is a method for the class \code{VECM}.
}
\details{
The \code{plot.VECM} function displays different types of plots based on the user's selection. The options are the following:

1: error correction terms \cr
2: raw residuals \cr
3: raw residuals standardized \cr
4: squared residuals \cr
5: squared residuals standardized \cr
6: autocorrelation functions of raw residuals \cr
7: autocorrelation functions of squared residuals \cr
8: partial autocorrelation functions of raw residuals \cr
9: partial autocorrelation functions of squared residuals \cr
10: kernel density estimations of raw residuals \cr

In plotting the error correction term, you can further choose the following: \cr

1: representation without deterministic terms \cr
2: representation with deterministic terms (if any) \cr
3: representation without deterministic terms and without short-run dynamics \cr
4: representation with deterministic terms (if any) but without short-run dynamics
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from Lütkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

plot(model)

}

}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
