% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VECM_forecast.R
\name{plot.VECM_forecast}
\alias{plot.VECM_forecast}
\title{Plotting VEC model forecasts}
\usage{
\method{plot}{VECM_forecast}(x, type = c("area", "interval", "line"), show = 50, observed = FALSE)
}
\arguments{
\item{x}{An object of class \code{VECM_forecast}.}

\item{type}{The type of plot. It can be chosen between \code{area} (the default), \code{interval}, and \code{line}.}

\item{show}{The number of observations for which the forecast is computed.}

\item{observed}{A logical value. If \code{TRUE} (not the default), the observed values are plotted together with the forecasts.}
}
\description{
This function is a method for the class \code{VECM_forecast}.
}
\seealso{
\code{\link{VECM_forecast}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
