% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VECM_IR.R
\name{plot.VECM_IR}
\alias{plot.VECM_IR}
\title{Plotting VEC model IRFs}
\usage{
\method{plot}{VECM_IR}(x, type = c("area", "line"), cum = FALSE)
}
\arguments{
\item{x}{An object of class \code{VECM_IR}.}

\item{type}{The type of plot. It can be chosen between \code{area} (the default), and \code{line}.}

\item{cum}{A logical value. If \code{TRUE} (not the default), accumulated responses are considered.}
}
\description{
This function is a method for the class \code{VECM_IR}.
}
\seealso{
\code{\link{VECM_IR}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
