% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_test.R
\name{port_test}
\alias{port_test}
\title{Portmanteau autocorrelation test}
\usage{
port_test(object, lag = 10)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{lag}{The number of tested lagged autocorrelations.}
}
\value{
An object of class \code{port_test} is a list with the following components: \cr

\item{stat}{The test statistic.}
\item{stat_pval}{The p-value of the test statistic.}
\item{stat_adj}{The adjusted test statistic.}
\item{stat_adj_pval}{The p-value of the adjusted test statistic.}
\item{lag}{The number of tested lagged autocorrelations.}
\item{df}{The degrees of freedom of the test statistic.}
}
\description{
This function tests the serial correlation of residuals using a portmanteau test.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)
test_p <- port_test(object = model, lag = 10)

print(test_p)
}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
