% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ARCH_test.R
\name{print.ARCH_test}
\alias{print.ARCH_test}
\title{Printing ARCH test}
\usage{
\method{print}{ARCH_test}(x)
}
\arguments{
\item{x}{An object of class \code{ARCH_test}.}
}
\description{
This function is a method for the class \code{ARCH_test}.
}
\seealso{
\code{\link{ARCH_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
