% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BG_test.R
\name{print.BG_test}
\alias{print.BG_test}
\title{Printing Breusch-Godfrey autocorrelation test}
\usage{
\method{print}{BG_test}(x)
}
\arguments{
\item{x}{An object of class \code{BG_test}.}
}
\description{
This function is a method for the class \code{BG_test}.
}
\seealso{
\code{\link{BG_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
