% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.chow_test.R
\name{print.chow_test}
\alias{print.chow_test}
\title{Printing Chow tests}
\usage{
\method{print}{chow_test}(x)
}
\arguments{
\item{x}{An object of class \code{chow_test}.}
}
\description{
This function is a method for the class \code{chow_test}.
}
\seealso{
\code{\link{chow_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
