% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cointegration_test.R
\name{print.cointegration_test}
\alias{print.cointegration_test}
\title{Printing cointegration test}
\usage{
\method{print}{cointegration_test}(x)
}
\arguments{
\item{x}{An object of class \code{cointegration_test}.}
}
\description{
This function is a method for the class \code{cointegration_test}.
}
\seealso{
\code{\link{cointegration_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
