% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.granger_test.R
\name{print.granger_test}
\alias{print.granger_test}
\title{Printing Granger-causality test}
\usage{
\method{print}{granger_test}(x)
}
\arguments{
\item{x}{An object of class \code{granger_test}.}
}
\description{
This function is a method for the class \code{granger_test}.
}
\seealso{
\code{\link{granger_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
