% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.normality_test.R
\name{print.normality_test}
\alias{print.normality_test}
\title{Printing multivariate normality tests}
\usage{
\method{print}{normality_test}(x)
}
\arguments{
\item{x}{An object of class \code{normality_test}.}
}
\description{
This function is a method for the class \code{normality_test}.
}
\seealso{
\code{\link{normality_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
