% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.port_test.R
\name{print.port_test}
\alias{print.port_test}
\title{Printing portmanteau autocorrelation test}
\usage{
\method{print}{port_test}(x)
}
\arguments{
\item{x}{An object of class \code{port_test}.}
}
\description{
This function is a method for the class \code{port_test}.
}
\seealso{
\code{\link{port_test}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
