% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VECM_FEVD.R
\name{print.VECM_FEVD}
\alias{print.VECM_FEVD}
\title{Printing VEC model FEVDs}
\usage{
\method{print}{VECM_FEVD}(x)
}
\arguments{
\item{x}{An object of class \code{VECM_FEVD}.}
}
\description{
This function is a method for the class \code{VECM_FEVD}.
}
\seealso{
\code{\link{VECM_FEVD}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
