% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VECM_forecast.R
\name{print.VECM_forecast}
\alias{print.VECM_forecast}
\title{Printing VEC model forecasts}
\usage{
\method{print}{VECM_forecast}(x)
}
\arguments{
\item{x}{An object of class \code{VECM_forecast}.}
}
\description{
This function is a method for the class \code{VECM_forecast}.
}
\seealso{
\code{\link{VECM_forecast}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
