% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VECM_IR.R
\name{print.VECM_IR}
\alias{print.VECM_IR}
\title{Printing impulse response functions}
\usage{
\method{print}{VECM_IR}(x, cum = FALSE)
}
\arguments{
\item{x}{An object of class \code{VECM_IR}.}

\item{cum}{A logical value. If \code{TRUE} (not the default), accumulated responses are considered.}
}
\description{
This function is a method for the class \code{VECM_IR}.
}
\seealso{
\code{\link{VECM_IR}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
