% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VECM_spec.R
\name{print.VECM_spec}
\alias{print.VECM_spec}
\title{Printing VEC model specification test}
\usage{
\method{print}{VECM_spec}(x)
}
\arguments{
\item{x}{An object of class \code{VECM_spec}.}
}
\description{
This function is a method for the class \code{VECM_spec}.
}
\seealso{
\code{\link{VECM_forecast}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
