% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_dummies.R
\name{seasonal_dummies}
\alias{seasonal_dummies}
\title{Seasonal dummy variables}
\usage{
seasonal_dummies(season, season_ref = NULL)
}
\arguments{
\item{season}{The number of seasons.}

\item{season_ref}{The season taken as reference.}
}
\value{
A matrix containing the seasonal dummy variables.
}
\description{
This function creates seasonal dummy variables.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
\keyword{internal}
