% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_check.R
\name{stability_check}
\alias{stability_check}
\title{Stability condition check}
\usage{
stability_check(object, plot = TRUE)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{plot}{A logical value. If \code{TRUE} (the default), a plot shows the roots of the companion matrix.}
}
\value{
A list containing the following components: \cr

\item{Re}{The real part of the roots of the companion matrix.}
\item{Im}{The complex part of the roots of the companion matrix.}
\item{Mod}{The modulus of the roots of the companion matrix.}
}
\description{
This function checks the eigenvalues stability condition in a VEC model.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)
check <- stability_check(model)
}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
