% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.VECM.R
\name{summary.VECM}
\alias{summary.VECM}
\title{Summarizing VEC model}
\usage{
\method{summary}{VECM}(x)
}
\arguments{
\item{x}{An object of class \code{VECM}.}
}
\description{
This function is a method for the class \code{VECM}.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
