% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM.R
\name{VECM}
\alias{VECM}
\title{Fit a VEC model}
\usage{
VECM(
  data,
  p,
  r,
  method = c("ML", "EGLS"),
  spec = c("none", "rconst", "uconst", "rtrend", "utrend"),
  shift = NULL,
  season = NULL,
  season_ref = NULL,
  exogen_ect = NULL,
  q_ect = NULL,
  exogen = NULL,
  q = NULL
)
}
\arguments{
\item{data}{The matrix containing the multivariate time series to be analyzed.}

\item{p}{The lag order. See below for more details.}

\item{r}{The cointegrating rank.}

\item{method}{The estimation method. It can be selected between \code{ML} (the default) and \code{EGLS}.}

\item{spec}{The type of deterministic component to be considered. See below for more details.}

\item{shift}{The break point at which there is a change in the levels of the variables. See below for more details.}

\item{season}{The number of seasonal dummy variables.}

\item{season_ref}{The season selected as reference. If \code{NULL}, the last season is automatically used as reference.}

\item{exogen_ect}{The exogenous variable to be included in the error correction term.}

\item{q_ect}{The lag order of the exogenous variable in the error correction term.}

\item{exogen}{The exogenous variable to be included outside the error correction term.}

\item{q}{The lag order of the exogenous variable outside the error correction term.}
}
\value{
An object of class \code{VECM} is a list with the following components: \cr

\item{alpha}{The loading matrix.}
\item{sigma_alpha}{The covariance matrix of \code{alpha}.}
\item{beta}{The cointegrating matrix.}
\item{sigma_beta}{The covariance matrix of \code{beta}.}
\item{pi}{The long-run impact matrix.}
\item{sigma_pi}{The covariance matrix of \code{pi}.}
\item{gamma}{The short-run impact matrix.}
\item{sigma_gamma}{The covariance matrix of \code{gamma}.}
\item{sigma_alphagamma}{The joint covariance matrix of \code{alpha} and \code{gamma}.}
\item{sigma_pigamma}{The joint covariance matrix of \code{pi} and \code{gamma}.}
\item{u}{The residuals matrix.}
\item{sigma_u}{The covariance matrix of \code{u}.}
\item{model}{A list containing information about the estimated model.}
}
\description{
This function fits a vector error correction model by using either the maximum likelihood (\code{ML}) or the estimated generalized least squares (\code{EGLS}) method.
}
\details{
The lag order (\code{p}) to which the user must specify refers to the lag in the VECM representation. It cannot be equal to zero.

The seasonal dummies are othogonal and their construction is based on  the description in Lütkepohl(2005).
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from Lütkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

print(model)
summary(model)
}

}
\references{
Johansen, S. (1995). \emph{Likelihood-based inference in cointegrated vector autoregressive models}. Oxford: Oxford University Press.

Lütkepohl, H., & Krätzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{plot.VECM}} \code{\link{print.VECM}} \code{\link{summary.VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
