% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_FEVD.R
\name{VECM_FEVD}
\alias{VECM_FEVD}
\title{Forecast error variance decomposition}
\usage{
VECM_FEVD(object, horizon = 10)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{horizon}{The time period over which the decomposition is performed.}
}
\value{
An object of class \code{VECM_FEVD} is a list with the following components: \cr

\item{W}{The percentage of the forecast variance due to each random innovation.}
\item{forecast_error}{Forecast standard errors.}
}
\description{
This function calculates the forecast error variance decomposition.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

fevd <- VECM_FEVD(object = model, horizon = 10)
print(fevd)
}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
