% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_forecast.R
\name{VECM_forecast}
\alias{VECM_forecast}
\title{Forecasting using VEC models}
\usage{
VECM_forecast(
  object,
  horizon = 10,
  type = c("static", "dynamic"),
  alpha = 0.05
)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{horizon}{The forecasting horizon.}

\item{type}{The type of forecast. It can be chosen between \code{static} (the default) and \code{dynamic}.}

\item{alpha}{The confidence level of the forecast confidence intervals.}
}
\value{
An object of class \code{VECM_forecast} is a list with the following components: \cr

\item{forecast}{The matrix containing the forecast of the levels variables.}
\item{sigma_forecast}{The covariance matrix of the matrix \code{forecast}.}
\item{CI_forecast}{The array containing the confidence intervals of the forecasts.}
\item{type}{The type of forecast.}
\item{alpha}{The confidence level of the forecast confidence intervals.}
\item{data}{The matrix containing the observed values.}
\item{u}{The residuals matrix.}
}
\description{
This function returns forecasts for VEC models.
}
\examples{
\dontrun{
#' require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

forecast <- VECM_forecast(object = model, type = "dynamic")

print(forecast)

plot(forecast)
}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
