% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_IR.R
\name{VECM_IR}
\alias{VECM_IR}
\title{Impulse responses analysis}
\usage{
VECM_IR(
  object,
  horizon = 10,
  transform = c("none", "orthogonal"),
  method = c("none", "analytic")
)
}
\arguments{
\item{object}{An object of class \code{VECM}.}

\item{horizon}{The forecasting horizon.}

\item{transform}{The residuals transformation. It can be chosen between \code{none} (the default) and \code{orthogonal}.}

\item{method}{The method used to compute IR standard errors. It can be chosen between \code{none} (the default) and \code{analytic}.}
}
\value{
An object of class \code{VECM_IR} is a list with the following components: \cr

\item{theta}{The matrix containing the estimated impulse response functions.}
\item{sigma_theta}{The covariance matrix of the matrix \code{theta}.}
\item{theta_cum}{The matrix containing the accumulated estimates of the impulse response functions.}
\item{sigma_theta_cum}{The covariance matrix of the matrix \code{theta_cum}.}
\item{transform}{The residuals transformation.}
\item{method}{The method used to compute IR standard errors.}
}
\description{
This function calculates impulse response functions for VEC models.
}
\examples{
\dontrun{
#' require(bvartools)

## The data for this example are taken from L체tkepohl, 2005.
data("e6")

model <- VECM(data = e6, p = 3,  r = 1, method = "EGLS", spec = "uconst", season = 4)

IR <- VECM_IR(object = model, transform = "orthogonal", method = "analytic"))

print(x = IR, cum = TRUE)

}

}
\references{
L체tkepohl, H., & Kr채tzig, M. (2004). \emph{Applied time series econometrics}. Cambridge: Cambridge University Press.

L체tkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
