% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_MA.R
\name{VECM_MA}
\alias{VECM_MA}
\title{Moving average impact matrix}
\usage{
VECM_MA(object)
}
\arguments{
\item{object}{An object of class \code{VECM}.}
}
\value{
A list containing the following components: \cr

\item{C}{The moving average impact matrix.}
\item{sigma_C}{The covariance matrix of the matrix \code{C}.}
}
\description{
This function provides the MA representation of the VEC model (i.e. Granger-representation theorem).
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
