% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_spec.R
\name{VECM_spec}
\alias{VECM_spec}
\title{VEC model specification test}
\usage{
VECM_spec(
  data,
  r,
  p,
  spec_H0 = c("none", "rconst", "uconst", "rtrend"),
  spec_H1 = c("rconst", "uconst", "rtrend", "utrend"),
  ...
)
}
\arguments{
\item{data}{A matrix containing the multivariate time series to be analyzed.}

\item{r}{The number of cointegrating relations.}

\item{p}{The lag order of the VEC model.}

\item{spec_H0}{The model with constrained deterministic terms. It can be chosen between \code{none} (the default), \code{rconst}, \code{uconst}, and \code{rtrend}.}

\item{spec_H1}{The model with no constrained deterministic terms. It can be chosen between \code{rconst} (the default), \code{uconst}, \code{rtrend}, and \code{utrend}.}

\item{...}{Further arguments passed to the \code{VECM} function.}
}
\value{
An object of class \code{VECM_spec} is a list with the following components: \cr

\item{spec_H0}{The model with constrained deterministic terms.}
\item{spec_H1}{The model with no constrained deterministic terms.}
\item{stat}{The test statistic.}
\item{df}{The degrees of freedom.}
\item{p_value}{The p-value of the test statistic.}
}
\description{
This function implements a likelihood ratio test for hypotheses regarding the deterministic terms of a VEC model.
}
\details{
The model need to be tested. For example, the user cannot test \code{spec_H0 = utrend} versus \code{spec_H1 = rtrend}.
}
\examples{
\dontrun{
require(bvartools)

## The data for this example are taken from Lütkepohl, 2005.
data("e6")

test_s <- VECM_spec(data = e6, r = 1, p = 3, spec_H0 = "uconst", spec_H1 = "rtrend")
print(test_s)
}

}
\references{
Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
