% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM_to_VAR.R
\name{VECM_to_VAR}
\alias{VECM_to_VAR}
\title{From VECM to VAR}
\usage{
VECM_to_VAR(object)
}
\arguments{
\item{object}{An object of class \code{VECM}.}
}
\value{
An object of class \code{VAR} with the following components: \cr

\item{A}{A matrix containing the VAR coefficients.}
\item{sigma_A}{The covariance matrix of the matrix \code{A}.}
\item{Z}{A matrix containing both the current levels and the corresponding lagged levels data.}
}
\description{
This function transforms a VEC model into a level VAR model.
}
\references{
Lütkepohl, H. (2005). \emph{New introduction to multiple time series analysis}. New York: Springer.
}
\seealso{
\code{\link{VECM}}
}
\author{
Gianluca Carpigo \email{gianluca.carpigo@uniroma1.it}

Michele Cianfriglia \email{michele.cianfriglia@libero.it}
}
